Refactor customer observer, remove address event
Change the customer save event to 'customer_save_before'.  Split the
delete and save functionality into separate classes.  Remove the
customer address save event listener.
