+++ etc/adminhtml/events.xml
@@ -38,13 +38,10 @@
     <event name="controller_action_predispatch">
         <observer name="taxjar_salestax_predispatch_admin_action_controller_observer" instance="Taxjar\SalesTax\Observer\AdminNotifications"/>
     </event>
-    <event name="customer_save_after_data_object">
-        <observer name="taxjar_salestax_customer_save_after_data_object" instance="Taxjar\SalesTax\Observer\Customer"/>
+    <event name="customer_save_before">
+        <observer name="taxjar_salestax_customer_save_before" instance="Taxjar\SalesTax\Observer\Customer\Save"/>
     </event>
     <event name="customer_delete_before">
-        <observer name="taxjar_salestax_customer_delete_before" instance="Taxjar\SalesTax\Observer\Customer"/>
-    </event>
-    <event name="customer_address_save_after">
-        <observer name="taxjar_salestax_customer_address_save_after" instance="Taxjar\SalesTax\Observer\Customer"/>
+        <observer name="taxjar_salestax_customer_delete_before" instance="Taxjar\SalesTax\Observer\Customer\Delete"/>
     </event>
 </config>
